FEAT (Header) Pull partials directly into header patter
Rather than creating an iframe to pull in the code for a rendered pattern example, the code should now be pulled directly into the page.

TODO: This needs some tidying to elegantly run the loop on each example (render the example, then display the code).
TODO: It would be nice if the code could be syntax highlighed.