FEAT (Code) Add documentation and examples
Rework pattern and options file so that it outputs only the HTML required to add a code block. The other examples are documentation and so should be added to the documentation page, not as options. Another factor for doing it this way is that JSON does not support multiple lines which are required for the code examples.